Move the BatchEnvironment service to require less deployment dependencies for the JBeret SPI implementation.
