+++ batch/extension-jberet/src/main/java/org/wildfly/extension/batch/jberet/impl/ClassLoaderContextHandle.java
@@ -1,0 +1,48 @@
+/*
+ * JBoss, Home of Professional Open Source.
+ * Copyright 2015, Red Hat, Inc., and individual contributors
+ * as indicated by the @author tags. See the copyright.txt file in the
+ * distribution for a full listing of individual contributors.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+ */
+
+package org.wildfly.extension.batch.jberet.impl;
+
+import org.wildfly.security.manager.WildFlySecurityManager;
+
+/**
+ * @author <a href="mailto:jperkins@redhat.com">James R. Perkins</a>
+ */
+class ClassLoaderContextHandle implements ContextHandle {
+    private final ClassLoader classLoader;
+
+    ClassLoaderContextHandle(final ClassLoader classLoader) {
+        this.classLoader = classLoader;
+    }
+
+    @Override
+    public Handle setup() {
+        final ClassLoader current = WildFlySecurityManager.getCurrentContextClassLoaderPrivileged();
+        WildFlySecurityManager.setCurrentContextClassLoaderPrivileged(classLoader);
+        return new Handle() {
+            @Override
+            public void tearDown() {
+                WildFlySecurityManager.setCurrentContextClassLoaderPrivileged(current);
+            }
+        };
+    }
+}
