+++ batch/extension-jberet/src/main/java/org/wildfly/extension/batch/jberet/impl/WildFlyArtifactFactory.java
@@ -1,6 +1,6 @@
 /*
  * JBoss, Home of Professional Open Source.
- * Copyright 2013, Red Hat, Inc., and individual contributors
+ * Copyright 2015, Red Hat, Inc., and individual contributors
  * as indicated by the @author tags. See the copyright.txt file in the
  * distribution for a full listing of individual contributors.
  *
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.jberet;
+package org.wildfly.extension.batch.jberet.impl;
 
 import java.util.Iterator;
 import java.util.Set;
