+++ feature-pack/src/main/resources/modules/system/layers/base/org/wildfly/extension/batch/jberet/main/module.xml
@@ -50,6 +50,7 @@
         <module name="org.jboss.logging"/>
         <module name="org.jboss.modules"/>
         <module name="org.jboss.msc"/>
+        <module name="org.picketbox"/>
         <module name="org.jboss.staxmapper"/>
         <module name="org.jboss.vfs"/>
         <module name="org.wildfly.extension.request-controller"/>
