+++ feature-pack/src/main/resources/modules/system/layers/base/org/wildfly/jberet/main/module.xml
@@ -36,12 +36,7 @@
         <module name="javaee.api"/>
         <module name="javax.transaction.api"/>
         <module name="org.jberet.jberet-core" services="import"/>
-        <module name="org.jboss.as.ee"/>
-        <module name="org.jboss.as.naming"/>
         <module name="org.jboss.logging"/>
-        <module name="org.jboss.msc"/>
-        <module name="org.picketbox"/>
-        <module name="org.wildfly.extension.request-controller"/>
         <module name="org.wildfly.security.elytron"/>
     </dependencies>
 </module>
